+++ src/Products/MeetingPROVHainaut/tests/testCustomWorkflows.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 
+from imio.helpers.cache import get_cachekey_volatile
 from Products.MeetingPROVHainaut.testing import MPH_FIN_TESTING_PROFILE_FUNCTIONAL
 from Products.MeetingPROVHainaut.tests.MeetingPROVHainautTestCase import MeetingPROVHainautTestCase
 from Products.MeetingPROVHainaut.utils import finance_group_cec_uid
@@ -16,8 +17,15 @@ class testCustomWorkflows(MeetingPROVHainautTestCase):
         """
            Test finances advices workflow.
         """
-        cfg = self.meetingConfig
+        def _check_date(item, modified_date, volatile_date):
+            '''Check that item modified date was updated.'''
+            new_modified_date = item.modified()
+            self.assertNotEqual(modified_date, new_modified_date)
+            new_volatile_date = get_cachekey_volatile('Products.PloneMeeting.MeetingItem.modified')
+            self.assertNotEqual(volatile_date, new_volatile_date)
+            return new_modified_date, new_volatile_date
 
+        cfg = self.meetingConfig
         self.changeUser('dgen')
         gic1_uid = cfg.getOrderedGroupsInCharge()[0]
         item = self.create('MeetingItem', groupsInCharge=(gic1_uid, ))
@@ -26,7 +34,7 @@ class testCustomWorkflows(MeetingPROVHainautTestCase):
         # ask finances advice
         fin_group_uid = finance_group_uid()
         item.setOptionalAdvisers((fin_group_uid + '__rowid__unique_id_002', ))
-        item.at_post_edit_script()
+        item._update_after_edit()
         # advice still not askable, askable as level2 or level3
         self.assertEqual(self.transitions(item),
                          ['proposeToValidationLevel1'])
@@ -49,13 +57,20 @@ class testCustomWorkflows(MeetingPROVHainautTestCase):
         # advice giveable when item complete
         self.assertFalse(item.adviceIndex[fin_group_uid]['advice_addable'])
         self.assertTrue(item.adapted().mayEvaluateCompleteness())
+        # we will check that item modified date is invalidated when advice changed
+        # this is responsible for updating collections counter in faceted portlet
+        volatile_date = get_cachekey_volatile('Products.PloneMeeting.MeetingItem.modified')
+        item_modified = item.modified()
         item.setCompleteness('completeness_complete')
-        item.at_post_edit_script()
+        item._update_after_edit()
+        item_modified, volatile_date = _check_date(item, item_modified, volatile_date)
         advice_portal_type = item._advicePortalTypeForAdviser(fin_group_uid)
         advice = self.addAdvice(item,
                                 advice_group=fin_group_uid,
                                 advice_type='positive_finance',
                                 advice_portal_type=advice_portal_type)
+        # item modified date was updated
+        item_modified, volatile_date = _check_date(item, item_modified, volatile_date)
         self.assertTrue(advice.advice_hide_during_redaction)
         self.assertEqual(self.transitions(advice),
                          ['proposeToFinancialController'])
@@ -70,6 +85,8 @@ class testCustomWorkflows(MeetingPROVHainautTestCase):
         # indexAdvisers is correctly reindexed
         advice_index_value = "delay__{0}_proposed_to_financial_controller".format(fin_group_uid)
         self.assertTrue(self.catalog(UID=item_uid, indexAdvisers=[advice_index_value]))
+        # item modified date was updated
+        item_modified, volatile_date = _check_date(item, item_modified, volatile_date)
         # financial editor
         self.do(advice, 'proposeToFinancialEditor')
         self.assertEqual(self.transitions(advice),
@@ -78,6 +95,8 @@ class testCustomWorkflows(MeetingPROVHainautTestCase):
         # indexAdvisers is correctly reindexed
         advice_index_value = "delay__{0}_proposed_to_financial_editor".format(fin_group_uid)
         self.assertTrue(self.catalog(UID=item_uid, indexAdvisers=[advice_index_value]))
+        # item modified date was updated
+        item_modified, volatile_date = _check_date(item, item_modified, volatile_date)
         # financial reviewer
         self.do(advice, 'proposeToFinancialReviewer')
         self.assertEqual(self.transitions(item), [])
@@ -88,6 +107,8 @@ class testCustomWorkflows(MeetingPROVHainautTestCase):
         # indexAdvisers is correctly reindexed
         advice_index_value = "delay__{0}_proposed_to_financial_reviewer".format(fin_group_uid)
         self.assertTrue(self.catalog(UID=item_uid, indexAdvisers=[advice_index_value]))
+        # item modified date was updated
+        item_modified, volatile_date = _check_date(item, item_modified, volatile_date)
         # financial manager
         self.do(advice, 'proposeToFinancialManager')
         self.assertEqual(self.transitions(item), [])
@@ -98,6 +119,8 @@ class testCustomWorkflows(MeetingPROVHainautTestCase):
         # indexAdvisers is correctly reindexed
         advice_index_value = "delay__{0}_proposed_to_financial_manager".format(fin_group_uid)
         self.assertTrue(self.catalog(UID=item_uid, indexAdvisers=[advice_index_value]))
+        # item modified date was updated
+        item_modified, volatile_date = _check_date(item, item_modified, volatile_date)
         # sign advice
         self.do(advice, 'signFinancialAdvice')
         self.assertEqual(self.transitions(item),
@@ -110,12 +133,16 @@ class testCustomWorkflows(MeetingPROVHainautTestCase):
         # indexAdvisers is correctly reindexed
         advice_index_value = "delay__{0}_financial_advice_signed".format(fin_group_uid)
         self.assertTrue(self.catalog(UID=item_uid, indexAdvisers=[advice_index_value]))
+        # item modified date was updated
+        item_modified, volatile_date = _check_date(item, item_modified, volatile_date)
         # validate item
         self.do(item, 'backTo_validated_from_waiting_advices')
         self.assertEqual(item.query_state(), 'validated')
         # indexAdvisers is correctly reindexed
         advice_index_value = "delay__{0}_advice_given".format(fin_group_uid)
         self.assertTrue(self.catalog(UID=item_uid, indexAdvisers=[advice_index_value]))
+        # item modified date was updated
+        item_modified, volatile_date = _check_date(item, item_modified, volatile_date)
 
     def test_CompletenessEvaluationAskedAgain(self):
         """When item is sent for second+ time to the finances,